requirements.txt
faiss_cpu==1.7.4
InstructorEmbedding==1.0.1
InstructorEmbedding==1.0.1
langchain==0.0.267
openai==0.27.8
PyPDF2==3.0.1
python-dotenv==1.0.0
